split start to main and finished to end package
