+++ ReplyAdaptabilityCodelab/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:theme="@style/Theme.Reply"
         tools:targetApi="31">
         <activity
-            android:name=".ui.start.MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Reply">
@@ -22,11 +22,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".ui.finish.FinishedActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.Reply" />
     </application>
 
 </manifest>
